refactor(convert): extract shared runnable-file loading & snapshot persistence
Lift the runnable-file loader and execution-snapshot helpers that were
duplicated between @deepnote/cli and @deepnote/mcp into @deepnote/convert,
and point both consumers at the shared implementations. The two copies had
drifted (error wrapping, return shapes, stale-executionCount handling); this
unifies them.

- convert: new load-runnable-file.ts (loadRunnableFile, isRunnableExtension,
  RUNNABLE_EXTENSIONS, LoadRunnableFileError) and
  snapshot/save-execution-snapshot.ts (saveExecutionSnapshot,
  mergeOutputsIntoFile, getSnapshotPath)
- generateSnapshotFilename now takes an object parameter (output byte-identical)
- cli format-converter.ts / output-persistence.ts become thin wrappers that
  keep CLI-only concerns (directory rejection, FileResolutionError, debug
  logging, IOutput[]-typed public surface)
- mcp execution.ts delegates to the shared helpers; local copies removed

Behavior is preserved for existing inputs, with two deliberate, documented
changes:
- mergeOutputsIntoFile now retains a stale executionCount (MCP semantics),
  pinned by a new test
- filesystem read errors keep their Node code (ENOENT/EISDIR) so MCP's error
  envelope is unchanged; only unsupported-extension/parse failures throw the
  code-less LoadRunnableFileError
